+++ config/bazar.php
@@ -133,6 +133,14 @@ return [
     'media' => [
         'disk' => 'public',
         'chunk_expiration' => 1440,
+        'conversion' => [
+            'default' => 'gd',
+            'drivers' => [
+                'gd' => [
+                    'quality' => 70,
+                ],
+            ],
+        ],
     ],
 
 ];