+++ src/Support/Facades/Conversion.php
@@ -2,15 +2,14 @@
 
 namespace Bazar\Support\Facades;
 
-use Bazar\Contracts\Repositories\ConversionRepository;
+use Bazar\Contracts\Conversion\Manager;
+use Closure;
 use Illuminate\Support\Facades\Facade;
 
 /**
- * @method static void register(string $name, \Closure $callback)
- * @method static void remove(string $name)
  * @method static \Bazar\Contracts\Models\Medium perform(\Bazar\Contracts\Models\Medium $medium)
  *
- * @see \Bazar\Contracts\Repositories\ConversionRepository
+ * @see \Bazar\Contracts\Conversion\Manager
  */
 class Conversion extends Facade
 {
@@ -21,6 +20,39 @@ class Conversion extends Facade
      */
     protected static function getFacadeAccessor(): string
     {
-        return ConversionRepository::class;
+        return Manager::class;
+    }
+
+    /**
+     * Register a new conversion.
+     *
+     * @param  string  $name
+     * @param  \Closure  $callback
+     * @return void
+     */
+    public static function register(string $name, Closure $callback): void
+    {
+        static::getFacadeRoot()->registerConversion($name, $callback);
+    }
+
+    /**
+     * Remove the given conversion.
+     *
+     * @param  string  $name
+     * @return void
+     */
+    public static function remove(string $name): void
+    {
+        static::getFacadeRoot()->removeConversion($name);
+    }
+
+    /**
+     * Get all the registered conversions.
+     *
+     * @return array
+     */
+    public static function all(): array
+    {
+        return static::getFacadeRoot()->getConversions();
     }
 }