+++ tests/Feature/CartDriverTest.php
@@ -2,6 +2,7 @@
 
 namespace Bazar\Tests\Feature;
 
+use Bazar\Cart\Checkout;
 use Bazar\Cart\CookieDriver;
 use Bazar\Cart\Manager;
 use Bazar\Database\Factories\ProductFactory;
@@ -9,28 +10,27 @@ use Bazar\Database\Factories\VariantFactory;
 use Bazar\Events\CartTouched;
 use Bazar\Models\Cart;
 use Bazar\Models\Shipping;
-use Bazar\Services\Checkout;
 use Bazar\Support\Facades\Cart as CartFacade;
 use Bazar\Tests\TestCase;
 use Illuminate\Support\Facades\Event;
 
 class CartDriverTest extends TestCase
 {
-    protected $cart, $product, $variant;
+    protected $manager, $product, $variant;
 
     public function setUp(): void
     {
         parent::setUp();
 
-        $this->cart = $this->app->make(Manager::class);
+        $this->manager = $this->app->make(Manager::class);
         $this->product = ProductFactory::new()->create(['prices' => ['usd' => ['default' => 100]]]);
         $this->variant = $this->product->variants()->save(VariantFactory::new()->make([
             'option' => ['Size' => 'S'],
             'prices' => ['usd' => ['default' => 150]],
         ]));
 
-        $this->cart->add($this->product, 2, ['option' => ['Size' => 'L']]);
-        $this->cart->add($this->product, 1, ['option' => ['Size' => 'S']]);
+        $this->manager->add($this->product, 2, ['option' => ['Size' => 'L']]);
+        $this->manager->add($this->product, 1, ['option' => ['Size' => 'S']]);
     }
 
     /** @test */
@@ -46,10 +46,10 @@ class CartDriverTest extends TestCase
     }
 
     /** @test */
-    public function it_can_be_retrieved_via_cookie_driver()
+    public function it_has_cart_driver()
     {
-        $this->assertInstanceOf(CookieDriver::class, $this->cart->driver('cookie'));
-        $this->assertInstanceOf(Cart::class, $this->cart->driver('cookie')->model());
+        $this->assertInstanceOf(CookieDriver::class, $this->manager->driver('cookie'));
+        $this->assertInstanceOf(Cart::class, $this->manager->driver('cookie')->model());
     }
 
     /** @test */
@@ -57,64 +57,64 @@ class CartDriverTest extends TestCase
     {
         Event::fake(CartTouched::class);
 
-        $this->cart->add($this->product, 2, ['option' => ['Size' => 'L']]);
+        $this->manager->add($this->product, 2, ['option' => ['Size' => 'L']]);
 
-        $this->assertEquals(5, $this->cart->count());
-        $this->assertEquals(2, $this->cart->items()->count());
-        $this->assertEquals(2, $this->cart->products()->count());
+        $this->assertEquals(5, $this->manager->count());
+        $this->assertEquals(2, $this->manager->items()->count());
+        $this->assertEquals(2, $this->manager->products()->count());
 
-        $product = $this->cart->item($this->product, ['option' => ['Size' => 'L']]);
+        $product = $this->manager->item($this->product, ['option' => ['Size' => 'L']]);
         $this->assertEquals(100, $product->price);
         $this->assertEquals(4, $product->quantity);
 
-        $variant = $this->cart->item($this->product, ['option' => ['Size' => 'S']]);
+        $variant = $this->manager->item($this->product, ['option' => ['Size' => 'S']]);
         $this->assertEquals(150, $variant->price);
         $this->assertEquals(1, $variant->quantity);
 
         Event::assertDispatched(CartTouched::class, function ($event) {
-            return $event->cart->id === $this->cart->model()->id;
+            return $event->cart->id === $this->manager->model()->id;
         });
     }
 
     /** @test */
     public function it_can_remove_items()
     {
-        $item = $this->cart->item($this->product, ['option' => ['Size' => 'L']]);
-        $this->cart->remove($item);
+        $item = $this->manager->item($this->product, ['option' => ['Size' => 'L']]);
+        $this->manager->remove($item);
 
-        $this->assertEquals(1, $this->cart->count());
-        $this->assertEquals(1, $this->cart->items()->count());
+        $this->assertEquals(1, $this->manager->count());
+        $this->assertEquals(1, $this->manager->items()->count());
     }
 
     /** @test */
     public function it_can_be_updated()
     {
-        $item = $this->cart->item($this->product, ['option' => ['Size' => 'L']]);
-        $this->cart->update([$item->id => ['quantity' => 10]]);
+        $item = $this->manager->item($this->product, ['option' => ['Size' => 'L']]);
+        $this->manager->update([$item->id => ['quantity' => 10]]);
 
-        $this->assertEquals(11, $this->cart->count());
-        $this->assertEquals(2, $this->cart->items()->count());
+        $this->assertEquals(11, $this->manager->count());
+        $this->assertEquals(2, $this->manager->items()->count());
     }
 
     /** @test */
     public function it_can_be_emptied()
     {
-        $this->assertTrue($this->cart->isNotEmpty());
-        $this->cart->empty();
-        $this->assertTrue($this->cart->isEmpty());
+        $this->assertTrue($this->manager->isNotEmpty());
+        $this->manager->empty();
+        $this->assertTrue($this->manager->isEmpty());
     }
 
     /** @test */
     public function it_has_shipping()
     {
-        $this->assertInstanceOf(Shipping::class, $this->cart->shipping());
+        $this->assertInstanceOf(Shipping::class, $this->manager->shipping());
     }
 
     /** @test */
     public function it_has_total()
     {
         $this->assertEquals(
-            $this->cart->model()->total, $this->cart->total()
+            $this->manager->model()->total, $this->manager->total()
         );
     }
 
@@ -122,7 +122,7 @@ class CartDriverTest extends TestCase
     public function it_has_tax()
     {
         $this->assertEquals(
-            $this->cart->model()->tax, $this->cart->tax()
+            $this->manager->model()->tax, $this->manager->tax()
         );
     }
 
@@ -130,7 +130,7 @@ class CartDriverTest extends TestCase
     public function it_has_discount()
     {
         $this->assertEquals(
-            $this->cart->model()->discount, $this->cart->discount()
+            $this->manager->model()->discount, $this->manager->discount()
         );
     }
 
@@ -138,7 +138,7 @@ class CartDriverTest extends TestCase
     public function it_can_checkout()
     {
         $this->assertInstanceOf(
-            Checkout::class, $this->cart->checkout()
+            Checkout::class, $this->manager->checkout()
         );
     }
 }